Refactor authentication and course management features
This commit includes changes to both frontend and backend. On the backend, modifications were made to the auth and course controllers, upload and verifyToken middlewares, and the Course model. On the frontend, changes were made to various components and pages to improve user experience and functionality. New hooks for admin and tutor authentication were added, and the useIsLoggedIn hook was removed. The atoms for course, token, and user state in the Recoil store were also updated or added. Overall, these changes aim to enhance the security, efficiency, and usability of the application.